Add TODO tree highlights,  logic to button window and refactor code
