+++ .vscode/settings.json
@@ -2,5 +2,66 @@
     "[python]": {
         "editor.defaultFormatter": "ms-python.python"
     },
-    "python.formatting.provider": "none"
+    "python.formatting.provider": "none",
+    "todo-tree.highlights.defaultHighlight": {
+        "icon": "alert",
+        "type": "text-and-comment",
+        "foreground": "#000000",
+        "background": "#FFFFFF",
+        "opacity": 80,
+        "iconColour": "#0000FF",
+        "gutterIcon": true,
+    },
+    "todo-tree.highlights.customHighlight": {
+        "TODO": {
+            "icon": "check",
+            "foreground": "#000000",
+            "background": "#fca139",
+            "iconColour": "#fca139",
+        },
+        "NOTE":{
+            "icon": "note",
+            "foreground": "#000000",
+            "background": "#a5a5a5",
+            "iconColour": "#808080",
+        },
+        "FIXME": {
+            "icon": "flame",
+            "foreground": "#000000",
+            "background": "#fdf909",
+            "iconColour": "#fdf909",
+        },
+        "BUG": {
+            "icon": "bug",
+            "foreground": "#000000",
+            "background": "#ff0000",
+            "iconColour": "#ff0000",
+        },
+        "HACK": {
+            "icon": "tools"
+        },
+        "[ ]": {
+            "icon": "issue-draft",
+            "foreground": "#000000",
+            "background": "#FFFFFF",
+            "iconColour": "#fdf909",
+        },
+        "[x]": {
+            "icon": "issue-closed",
+            "foreground": "#000000",
+            "background": "#00FF00",
+            "iconColour": "#00FF00",
+        }
+    },
+    "todo-tree.general.tags": [
+        "TODO",
+        "NOTE",
+        "BUG",
+        "HACK",
+        "FIXME",
+        "XXX",
+        "[ ]",
+        "[x]"
+    ],
+    "todo-tree.regex.regex": "(//|#|<!--|;|/\\*|^|^\\s*-\\s*)\\s*($TAGS)|($TAGS)",
 }
